Rearrange commonly used functions of JMS Actions
