+++ component/src/main/ballerina/ballerina/net/jms/natives.bal
@@ -59,7 +59,7 @@ public connector JmsClient (ClientProperties clientProperties) {
     @Description {value:"POLL action implementation of the JMS Connector"}
     @Param {value:"destinationName: Destination Name"}
     @Param {value:"time: Timeout that needs to blocked on"}
-    native action poll (string destinationName, int time) (JMSMessage, error);
+    native action poll (string destinationName, int time) (JMSMessage);
 
 }
