+++ component/src/main/java/org/ballerinalang/net/jms/actions/Send.java
@@ -16,8 +16,6 @@
 
 package org.ballerinalang.net.jms.actions;
 
-import org.ballerinalang.bre.BallerinaTransactionContext;
-import org.ballerinalang.bre.BallerinaTransactionManager;
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.connector.api.ConnectorFuture;
 import org.ballerinalang.model.types.TypeKind;
@@ -28,11 +26,8 @@ import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.Attribute;
 import org.ballerinalang.natives.annotations.BallerinaAction;
 import org.ballerinalang.natives.annotations.BallerinaAnnotation;
-import org.ballerinalang.natives.annotations.ReturnType;
 import org.ballerinalang.net.jms.Constants;
-import org.ballerinalang.net.jms.JMSTransactionContext;
 import org.ballerinalang.net.jms.JMSUtils;
-import org.ballerinalang.util.DistributedTxManagerProvider;
 import org.ballerinalang.util.exceptions.BallerinaException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -45,28 +40,23 @@ import org.wso2.transport.jms.utils.JMSConstants;
 import java.util.Map;
 import java.util.UUID;
 import javax.jms.Message;
-import javax.transaction.TransactionManager;
 
 import static org.ballerinalang.net.jms.Constants.EMPTY_CONNECTOR_ID;
 
 /**
- * {@code Post} is the send action implementation of the JMS Connector.
+ * {@code Send} is the send action implementation of the JMS Connector.
  */
 @BallerinaAction(packageName = "ballerina.net.jms",
                  actionName = "send",
                  connectorName = Constants.CONNECTOR_NAME,
-                 args = {
-                         @Argument(name = "jmsClientConnector",
-                                   type = TypeKind.CONNECTOR), @Argument(name = "destinationName",
-                                                                         type = TypeKind.STRING),
-                         @Argument(name = "msgType",
-                                   type = TypeKind.STRING), @Argument(name = "m",
-                                                                      type = TypeKind.MESSAGE)
+                 args = {@Argument(name = "destinationName",
+                                   type = TypeKind.STRING),
+                         @Argument(name = "message",
+                                   type = TypeKind.STRUCT)
                  },
-                 returnType = { @ReturnType(type = TypeKind.BOOLEAN) },
                  connectorArgs = {
                          @Argument(name = "properties",
-                                   type = TypeKind.MAP)
+                                   type = TypeKind.STRUCT)
                  })
 @BallerinaAnnotation(annotationName = "Description",
                      attributes = {
@@ -75,23 +65,13 @@ import static org.ballerinalang.net.jms.Constants.EMPTY_CONNECTOR_ID;
                      })
 @BallerinaAnnotation(annotationName = "Param",
                      attributes = {
-                             @Attribute(name = "connector",
-                                        value = "Connector")
-                     })
-@BallerinaAnnotation(annotationName = "Param",
-                     attributes = {
                              @Attribute(name = "destinationName",
                                         value = "Destination Name")
                      })
 @BallerinaAnnotation(annotationName = "Param",
                      attributes = {
-                             @Attribute(name = "msgType",
-                                        value = "Message Type")
-                     })
-@BallerinaAnnotation(annotationName = "Param",
-                     attributes = {
                              @Attribute(name = "message",
-                                        value = "Message")
+                                        value = "JMS Message")
                      })
 public class Send extends AbstractJMSAction {
     private static final Logger log = LoggerFactory.getLogger(Send.class);
@@ -132,6 +112,8 @@ public class Send extends AbstractJMSAction {
         }
 
         try {
+            /* todo: Cache this created JMSClientConnector. Creating it per-request is costly because
+               todo: createClientConnector method of JMSConnectorFactoryImpl is having a synchronized call */
             JMSClientConnector jmsClientConnector = new JMSConnectorFactoryImpl().createClientConnector(propertyMap);
             if (log.isDebugEnabled()) {
                 log.debug("Sending JMS Message to " + propertyMap.get(JMSConstants.PARAM_DESTINATION_NAME));
@@ -139,23 +121,7 @@ public class Send extends AbstractJMSAction {
             if (!isTransacted) {
                 jmsClientConnector.send(jmsMessage, destination);
             } else {
-                SessionWrapper sessionWrapper;
-                BallerinaTransactionManager ballerinaTxManager = context.getBallerinaTransactionManager();
-                BallerinaTransactionContext txContext = ballerinaTxManager.getTransactionContext(connectorKey);
-                // if transaction initialization has not yet been done
-                // (if this is the first transacted action happens from this particular connector with this
-                // transaction block)
-                if (txContext == null) {
-                    sessionWrapper = jmsClientConnector.acquireSession();
-                    txContext = new JMSTransactionContext(sessionWrapper, jmsClientConnector);
-                    //Handle XA initialization
-                    if (txContext.getXAResource() != null) {
-                        initializeXATransaction(ballerinaTxManager);
-                    }
-                    ballerinaTxManager.registerTransactionContext(connectorKey, txContext);
-                } else {
-                    sessionWrapper = ((JMSTransactionContext) txContext).getSessionWrapper();
-                }
+                SessionWrapper sessionWrapper = getTxSession(context, jmsClientConnector, connectorKey);
                 jmsClientConnector.sendTransactedMessage(jmsMessage, destination, sessionWrapper);
             }
         } catch (JMSConnectorException e) {
@@ -164,16 +130,5 @@ public class Send extends AbstractJMSAction {
         ClientConnectorFuture future = new ClientConnectorFuture();
         future.notifySuccess();
         return future;
-    }
-
-    private void initializeXATransaction(BallerinaTransactionManager ballerinaTxManager) {
-        /* Atomikos transaction manager initialize only distributed transaction is present.*/
-        if (!ballerinaTxManager.hasXATransactionManager()) {
-            TransactionManager transactionManager = DistributedTxManagerProvider.getInstance().getTransactionManager();
-            ballerinaTxManager.setXATransactionManager(transactionManager);
-        }
-        if (!ballerinaTxManager.isInXATransaction()) {
-            ballerinaTxManager.beginXATransaction();
-        }
     }
 }
